Project-WebHealth, API-Testing/AbdullahZaman/sprint3/PCRepoAZB/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.aws-lambda==1.135.0
aws-cdk.core==1.135.0
aws-cdk.aws-lambda==1.135.0
aws_cdk.aws_events_targets==1.135.0
aws_cdk.aws_cloudwatch_actions==1.135.0
aws-cdk.aws_dynamodb==1.135.0
aws_cdk.pipelines==1.135.0
aws_cdk.aws_codepipeline_actions==1.135.0
aws_cdk.aws_codedeploy==1.135.0
aws_cdk.assertions==1.135.0
aws_cdk.aws_apigateway==1.135.0
boto3
pytest